+++ crates/wdl-engine/src/backend/slurm_apptainer.rs
@@ -802,11 +802,16 @@ impl SlurmApptainerBackend {
                 .unwrap_or(DEFAULT_MAX_CONCURRENCY) as usize,
         );
 
+        let apptainer = ApptainerRuntime::new(
+            run_root_dir,
+            backend_config.apptainer_config.image_cache_dir.as_deref(),
+        )?;
+
         Ok(Self {
             config,
             events,
             cancellation,
-            apptainer: ApptainerRuntime::new(run_root_dir)?,
+            apptainer,
             monitor,
             permits,
         })